==> Bootcamp/03_Machine_Learning/7-Recommendation_Systems/Teoria/0-Recomendador_Conocimiento.ipynb <==
# Sistema de Recomendación Basado en Conocimiento

import pandas as pd
import numpy as np

import warnings

warnings.filterwarnings('ignore')

df = pd.read_csv('data/movies_metadata.csv')
df.columns

Seleccionamos la variables que necesitamos:
* `title` 
* `genres` 
* `release_date` 
* `runtime`
* `vote_average` 
* `vote_count`

df = df[['title','genres', 'release_date', 'runtime', 'vote_average', 'vote_count']]
display(df.head(), df.info())

Convertimos la variable `release_date` a formato ***datetime*** y extraemos el año de la fecha.

df['release_date'] = pd.to_datetime(df['release_date'], errors='coerce')

df['year'] = df['release_date'].apply(lambda x: str(x).split('-')[0] if x != np.nan else np.nan)

df.head()

Función auxiliar para convertir ***NaT*** a 0 y todos los años a enteros.

def convert_int(x):
    try:
        return int(x)
    except:
        return 0

Aplicamos ***convert_int*** a la variable `year`.

df['year'] = df['year'].apply(convert_int)

Eliminamos la columna `release_date`.

df = df.drop('release_date', axis=1)
df.head()

Veamos los géneros de la primera película.

df.iloc[0]['genres']

Importamos la función `literal_eval` del módulo ***ast***.

from ast import literal_eval #nos permite evaluar el tipo de variable que es aunque este dentro de un string

a = "[1,2,3]"
print(type(a))

b = literal_eval(a)
print(type(b))

1. Convertimos todos los ***strings*** de listas vacías en ***NaN***.
1. Convertimos todos los valores a listas.
1. Convertimos la lista de diccionarios en una lista de ***strings***.

df['genres'] = df['genres'].fillna('[]')

df['genres'] = df['genres'].apply(literal_eval)

df['genres'] = df['genres'].apply(lambda x: [i['name'].lower() for i in x] if isinstance(x, list) else [])

df.head()

1. Creamos una nueva variable que separa los géneros de las películas.
1. Nombramos la nueva variable como `genre`.
1. Creamos un nuevo ***Data Frame*** eliminando la variable `genres`.

s = df.apply(lambda x: pd.Series(x['genres']),axis=1).stack().reset_index(level=1, drop=True)

s.name = 'genre'

gen_df = df.drop('genres', axis=1).join(s)

gen_df.head()

Definimos la función basada en conocimiento (cuestiones realizadas al usuario).

def build_chart(gen_df, percentile=0.8):
    
    genre = input("Introduce tu género favorito, ")
    
    low_time = int(input("Introduce la duración más corta deseada, "))
    
    high_time = int(input("Introduce la duración más larga deseada, "))
    
    low_year = int(input("Introduce el año más antiguo, "))
    
    high_year = int(input("Introduce el año más reciente, "))
    
    movies = gen_df.copy()
    movies = movies[(movies['genre'] == genre) & 
                    (movies['runtime'] >= low_time) & 
                    (movies['runtime'] <= high_time) & 
                    (movies['year'] >= low_year) & 
                    (movies['year'] <= high_year)]
    
    
    C = movies['vote_average'].mean()
    m = movies['vote_count'].quantile(percentile)
    
    # Solo consideramos aquellas películas con más de m votos
    q_movies = movies.copy().loc[movies['vote_count'] >= m]
    
    # Calculamos el score aplicando la fórmula de IMDB
    q_movies['score'] = q_movies.apply(lambda x: (x['vote_count']/(x['vote_count']+m) * x['vote_average']) 
                                       + (m/(m+x['vote_count']) * C)
                                       ,axis=1)

    # Ordenamos las películas de manera descendente segun su respectivo score
    q_movies = q_movies.sort_values('score', ascending=False)
    
    return q_movies

Generamos la lista del ***top*** 10 segun las preferencias del cliente.

build_chart(gen_df).head(10)

Guardamos nuestro ***Data Frame*** limpio.

df.to_csv('data/metadata_clean.csv', index=False)
