==> Bootcamp/03_Machine_Learning/7-Recommendation_Systems/Teoria/1-Recomendador_Simple.ipynb <==
# Sistema de Recomendación Simple

import pandas as pd
import numpy as np

import warnings
warnings.filterwarnings('ignore')

Cargamos el ***data set*** de las películas. 

df = pd.read_csv('data/movies_metadata.csv')
df.head()

Calculamos el número de votos obtenido por el percentil 80.

m = df['vote_count'].quantile(0.80)
m

Solo consideramos las películas con una duración superior o igual a 45 minutos y una duración inferior a 300 minutos, y que hayan obtenido más de $m$ votos.

q_movies = df[(df['runtime'] >= 45) & (df['runtime'] <= 300)]

q_movies = q_movies[q_movies['vote_count'] >= m]

q_movies.shape

Cálculo del valor $C$.

C = df['vote_average'].mean()
C

Función que calcula el ratio ponderado de IMDB para cada película.

def weighted_rating(x, m=m, C=C):
    v = x['vote_count']
    R = x['vote_average']
    
    return (v/(v+m) * R) + (m/(m+v) * C)

Cálculo del ***score*** usando la función del ratio ponderado definido previamente.

q_movies['score'] = q_movies.apply(weighted_rating, axis=1)

Ordenamos las películas en orden descendente por el ***score*** y mostramos el ***top*** 25.

q_movies = q_movies.sort_values('score', ascending=False)

q_movies[['title', 'vote_count', 'vote_average', 'score', 'runtime']].head(25)
