==> Bootcamp/03_Machine_Learning/7-Recommendation_Systems/Teoria/2-Recomendador_Basado_Contenido.ipynb <==
# Sistema de Recomendación Basado en Contenido

import pandas as pd
import numpy as np

import warnings
warnings.filterwarnings('ignore')

Cargamos los datos limpios de las películas.

df = pd.read_csv('data/metadata_clean.csv')

df.head()

Cargamos los datos originales de las películas y añadimos dos variables al conjunto de datos limpio.

orig_df = pd.read_csv('data/movies_metadata.csv', low_memory=False)

df['overview'], df['id'] = orig_df['overview'], orig_df['id']

df.head()

1. Importamos las librerías necesarias para vectorizar texto. 
2. Eliminamos las ***stopwords*** del inglés y sustituimos los NaN. 
3. Finalmente, aplicamos el algoritmo de vectorización de texto.

from sklearn.feature_extraction.text import TfidfVectorizer

tfidf = TfidfVectorizer(stop_words='english')

df['overview'] = df['overview'].fillna('')

tfidf_matrix = tfidf.fit_transform(df['overview'])

tfidf_matrix.shape

1. Importamos el kernel lineal para calcular el producto escalar.
2. Calculamos la matriz de similitud del coseno.

from sklearn.metrics.pairwise import linear_kernel

cosine_sim = linear_kernel(tfidf_matrix, tfidf_matrix)

Construir una correspondencia inversa de índices y títulos de películas, y eliminar los títulos duplicados, si los hay.

indices = pd.Series(df.index, index=df['title']).drop_duplicates()

Desarrollamos la función que toma el título de la película como entrada y da recomendaciones 

Funcionamiento:
1. Obtener el índice de la película que coincide con el título
1. Obtener las puntuaciones de similitud pairwsie de todas las películas con esa película y convertirlo en una lista de tuplas como se ha descrito anteriormente
1. Ordenar las películas en función de las puntuaciones de similitud coseno.
1. Obtener las puntuaciones de las 10 películas más similares. Ignorar la primera película.
1. Obtener los índices de las películas
1. Devuelve las 10 películas más parecidas

def content_recommender(title, cosine_sim=cosine_sim, df=df, indices=indices):
    
    idx = indices[title]

    sim_scores = list(enumerate(cosine_sim[idx]))

    sim_scores = sorted(sim_scores, key=lambda x: x[1], reverse=True)

    sim_scores = sim_scores[1:11]

    movie_indices = [i[0] for i in sim_scores]

    return df['title'].iloc[movie_indices]

Obtenemos la recomendación.

content_recommender('Jumanji')
