==> Bootcamp/03_Machine_Learning/7-Recommendation_Systems/Teoria/4-Filtrado_Colaborativo.ipynb <==
# Filtrado Colaborativo

import pandas as pd
import numpy as np

import warnings
warnings.filterwarnings('ignore')

Cargamos los datos de los usuarios.

u_cols = ['user_id', 'age', 'sex', 'occupation', 'zip_code']

users = pd.read_csv('data/ml-100k/u.user', sep='|', names=u_cols,
 encoding='latin-1')

display(users.shape, users.head())

Cargamos los datos de los items (productos, es decir, películas).

i_cols = ['movie_id', 'title' ,'release date','video release date', 'IMDb URL', 'unknown', 'Action', 'Adventure',
 'Animation', 'Children\'s', 'Comedy', 'Crime', 'Documentary', 'Drama', 'Fantasy',
 'Film-Noir', 'Horror', 'Musical', 'Mystery', 'Romance', 'Sci-Fi', 'Thriller', 'War', 'Western']

movies = pd.read_csv('data/ml-100k/u.item', sep='|', names=i_cols, encoding='latin-1')

display(movies.shape, movies.head())

Eliminamos toda la información de las películas, salvo el valor id e la película y el título.

movies = movies[['movie_id', 'title']]

Cargamos los datos sobre la relación entre usuarios y películas (100K registros).

r_cols = ['user_id', 'movie_id', 'rating', 'timestamp']

ratings = pd.read_csv('data/ml-100k/u.data', sep='\t', names=r_cols,
 encoding='latin-1')

display(ratings.shape, ratings.head())

Eliminamos la columna `timestamp`.

ratings = ratings.drop('timestamp', axis=1)

Dividimos de datos entre ***train*** y ***test***.

from sklearn.model_selection import train_test_split

X = ratings.copy()
y = ratings['user_id']

X_train, X_test, y_train, y_test = train_test_split(X, y, test_size = 0.25, stratify=y, random_state=42)

Importamos un criterio de evaluación. En este caso, el error cuadrático medio.

from sklearn.metrics import mean_squared_error

def rmse(y_true, y_pred):
    return np.sqrt(mean_squared_error(y_true, y_pred))

Definimos el modelo base para que siempre devuelva 3.

def baseline(user_id, movie_id):
    return 3.0

Desarrollamos la función para calcular la puntuación *RMSE* obtenida en el conjunto *test*.

1. Construir una lista de tuplas usuario-película a partir del conjunto de datos de prueba.
1. Predecir la puntuación de cada tupla usuario-película.
1. Extraer las valoraciones reales dadas por los usuarios en los datos de *test*.
1. Obtener el valor de *RMSE*.

def score(cf_model):
    
    id_pairs = zip(X_test['user_id'], X_test['movie_id'])
    
    y_pred = np.array([cf_model(user, movie) for (user, movie) in id_pairs])
    
    y_true = np.array(X_test['rating'])
    
    return rmse(y_true, y_pred)

score(baseline)

## Filtrado Colaborativo Basado en el Usuario

### Matriz de *Ratings*

Construimos la matriz de *ratings* utilizando la función `pivot_table`.

r_matrix = X_train.pivot_table(values='rating', index='user_id', columns='movie_id')

r_matrix.head()

### Media

Filtrado colaborativo basado en el usuario utilizando los *ratings* medios

1. Comprueba si `movie_id` existe en `r_matrix`.
1. Calcular la media de todas las valoraciones dadas a la película.
1. Por defecto, si no hay información, la calificación es 3.

def cf_user_mean(user_id, movie_id):
    
    if movie_id in r_matrix:
        mean_rating = r_matrix[movie_id].mean()
    
    else:
        mean_rating = 3.0
    
    return mean_rating

Calculamos el *RMSE* para el modelo *Mean*

score(cf_user_mean)

### Media Ponderada

Creamos una matriz de *ratings* ficticios con todos los valores nulos imputados a 0.

r_matrix_dummy = r_matrix.copy().fillna(0)

Importamos el método `cosine similarity` del submódulo ***pairwise*** y calculamos la matriz de la similitud del coseno sobre la matriz de *ratings* ficticios.

from sklearn.metrics.pairwise import cosine_similarity

cosine_sim = cosine_similarity(r_matrix_dummy, r_matrix_dummy)

Lo convertimos a un *Data Frame* de Pandas.

cosine_sim = pd.DataFrame(cosine_sim, index=r_matrix.index, columns=r_matrix.index)

cosine_sim.head(10)

Filtrado Colaborativo basado en el usuario utilizando *ratings* medios ponderados.

1. Comprueba si `movie_id` existe en `r_matrix`.
1. Obtiene las puntuaciones de similitud del usuario en cuestión con todos los demás usuarios.
1. Obtiene los *ratings* de los usuarios para la película en cuestión.
1. Extrae los índices que contienen ***NaN*** en la serie `m_ratings`.
1. Elimina los valores ***NaN*** de la serie `m_ratings`.
1. Elimina las puntuaciones de cosenos correspondientes de la serie `sim_scores`.
1. Calcula la media ponderada final.  
1. Valor predeterminado a una calificación de 3 en ausencia de información.

def cf_user_wmean(user_id, movie_id):
    
    if movie_id in r_matrix:
        
        sim_scores = cosine_sim[user_id]
        
        m_ratings = r_matrix[movie_id]
        
        idx = m_ratings[m_ratings.isnull()].index
        
        m_ratings = m_ratings.dropna()
        
        sim_scores = sim_scores.drop(idx)
        
        wmean_rating = np.dot(sim_scores, m_ratings)/ sim_scores.sum()
    
    else:
        
        wmean_rating = 3.0
    
    return wmean_rating

def score(cf_model):
    
    id_pairs = zip(X_test['user_id'], X_test['movie_id'])
    
    y_pred = np.array([cf_model(user, movie) for (user, movie) in id_pairs])
    
    y_true = np.array(X_test['rating'])
    print(y_pred, y_true)

    list_pred = list(np.argwhere(np.isnan(y_pred)))
    list_true = list(np.argwhere(np.isnan(y_true)))
    
    if len(list_pred)>0:
        list_pred = list(list_pred[0])
    if len(list_true)>0:
        list_true = list(list_true[0])
    
    list_res = list_pred + list_true
    print(list_res)
    return rmse(y_true, y_pred)

score(cf_user_wmean)

def score(cf_model):
    id_pairs = zip(X_test['user_id'], X_test['movie_id'])
    y_pred = np.array([cf_model(user, movie) for (user, movie) in id_pairs])
    y_pred[np.isnan(y_pred)] = False
    y_true = np.array(X_test['rating'])
    print(y_pred, y_true)
    list_pred = list(np.argwhere(np.isnan(y_pred)))
    list_true = list(np.argwhere(np.isnan(y_true)))
    if len(list_pred)>0:
        list_pred = list(list_pred[0])
    if len(list_true)>0:
        list_true = list(list_true[0])
    list_res = list_pred + list_true
    print(list_res)
    return rmse(y_true, y_pred)
import numpy as np

score(cf_user_wmean)

Unimos el *Data Frame* de los usuarios originales con el conjunto de *train*.

merged_df = pd.merge(X_train, users)

merged_df.head()

Calculamos el *rating* medio de cada película por género.

gender_mean = merged_df[['movie_id', 'sex', 'rating']].groupby(['movie_id', 'sex'])['rating'].mean()

Establecemos el índice del *Data Frame* de los usuarios en `user_id`.

users = users.set_index('user_id')

Filtrado colaborativo basado en género utilizando *ratings* medios.

1. Comprueba si `movie_id` existe en `r_matrix`.
1. Identifica el género del usuario.
1. Comprueba si el género ha calificado la película.
1. Calcula el *rating* medio otorgado por ese género a la película.
1. Valor predeterminado a una calificación de 3 en ausencia de información.

def cf_gender(user_id, movie_id):
    
    if movie_id in r_matrix:
        
        gender = users.loc[user_id]['sex']

        if gender in gender_mean[movie_id]:
            
            gender_rating = gender_mean[movie_id][gender]
        
        else:
            gender_rating = 3.0
    
    else:
        
        gender_rating = 3.0
    
    return gender_rating

score(cf_gender)

Calculamos el *rating* medio por género y ocupación.

gen_occ_mean = merged_df[['sex', 'rating', 'movie_id', 'occupation']].pivot_table(
    values='rating', index='movie_id', columns=['occupation', 'sex'], aggfunc='mean')

gen_occ_mean.head()

Filtrado colaborativo basado en género y ocupación utilizando *ratings* medios.
    
1. Comprueba si `movie_id` existe en `gen_occ_mean`.
1. Identifica al usuario.
1. Identifica el género y la ocupación.
1. Comprueba si la ocupación ha calificado la película.
1. Comprueba si el género ha calificado la película.
1. Extrae el *rating* requerido.
1. Predeterminado a 3 si el *rating* es un *NaN*.
1. Devuelve la calificación predeterminada.   

def cf_gen_occ(user_id, movie_id):
    
    if movie_id in gen_occ_mean.index:
        
        user = users.loc[user_id]
        
        gender = user['sex']
        occ = user['occupation']
        
        if occ in gen_occ_mean.loc[movie_id]:
            
            if gender in gen_occ_mean.loc[movie_id][occ]:
                
                rating = gen_occ_mean.loc[movie_id][occ][gender]
                
                if np.isnan(rating):
                    rating = 3.0
                
                return rating
              
    return 3.0

score(cf_gen_occ)

## Enfoques Basados en Modelos

Importamos las clases y métodos requeridos de la librería Surprise.

* El objeto `Reader` nos ayuda a analizar el archivo o *Data Frame* que contiene los *ratings*.
* Crea el conjunto de datos que se utilizará para crear el filtro.
* Definimos el objeto del algoritmo `kNN`.
* Evalua el modelo.

from surprise import Reader, Dataset, KNNBasic
from surprise.model_selection import cross_validate

reader = Reader()

data = Dataset.load_from_df(ratings, reader)

knn = KNNBasic()

cross_validate(knn, data, measures=['RMSE'], cv=5, verbose=True)

Importamos el algoritmo que realiza la descomposición de valores singulares (SVD, Singular Value Descomposition), lo aplicamos sobre nuestros datos y lo evaluamos.

from surprise import SVD

svd = SVD()

cross_validate(svd, data, measures=['RMSE'], cv=5, verbose=True)
